# internal/gen/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Judex API
  version: 1.0.0
  summary: Judex 协作平台公共 API（docs/plans/v1/06 契约源）
  description: |
    所有业务路径前缀 /api/v1（本规范路径不含该前缀，由网关/服务器统一挂载）。

    协议规则：JSON 字段 camelCase；UUID 字符串；时间 RFC3339 UTC；版本为正整数。
    成功 envelope `{data, meta:{requestId,serverTime,eventCursor?}}`；命令额外返回 `affected`。
    错误 envelope `{error:{code,message,details?,retryable},requestId}`；客户端按 code+details 处理，不解析 message。
    分页 `?limit=50&cursor=<opaque>`（1-100）；列表 data 为 `{items,nextCursor}`。
    有副作用的业务命令要求 Idempotency-Key 头（UUID）；已有对象写操作要求 expectedVersion；正式决定要求 reviewId/reviewHash。
    Web 会话使用 Cookie+CSRF（sessionCookie）；CLI 使用受限 Bearer grant（bearerGrant）。
    阶段实现状态由 GET /system capabilities 返回；未实现的 operation 返回 501 而不是假成功。
servers:
  - url: /api/v1
    description: Same-origin (SPA or CLI with explicit base URL)
...
security:
  - sessionCookie: []
  - bearerGrant: []
paths:
...
  /projects/{projectId}/confirmation-intents:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    post:
      operationId: createConfirmationIntent
      summary: CLI 创建人工确认意图（返回 confirmUrl，不是批准）
      tags:
        - intents
      security:
        - bearerGrant: []
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmationIntentRequest'
      responses:
        '201':
          description: 意图已创建。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ConfirmationIntent'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/confirmation-intents/{intentId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: intentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getConfirmationIntent
      summary: 查询意图状态（本人/grant；CLI 轮询用）
      tags:
        - intents
      security:
        - bearerGrant: []
        - sessionCookie: []
      responses:
        '200':
          description: 意图状态与 resultRef。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ConfirmationIntent'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/confirmation-intents/{intentId}/confirm:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: intentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: confirmIntent
      summary: 浏览器一次确认（仅 Web 会话；原子执行原领域命令）
      tags:
        - intents
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmIntentRequest'
      responses:
        '200':
          description: 确认结果（committed/rejected）与 resultRef。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ConfirmationIntent'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /confirmation-intents:
    parameters: []
    post:
      operationId: createGlobalConfirmationIntent
      summary: 用户级建项目确认意图（仅 project.create）
      tags:
        - intents
      security:
        - bearerGrant: []
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmationIntentRequest'
      responses:
        '201':
          description: 意图已创建。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ConfirmationIntent'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /confirmation-intents/{intentId}:
    parameters:
      - name: intentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getGlobalConfirmationIntent
      summary: 用户级建项目确认意图（仅 project.create）
      tags:
        - intents
      security:
        - bearerGrant: []
        - sessionCookie: []
      responses:
        '200':
          description: 意图状态与 resultRef。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ConfirmationIntent'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /confirmation-intents/{intentId}/confirm:
    parameters:
      - name: intentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: confirmGlobalIntent
      summary: 用户级建项目确认意图（仅 project.create）
      tags:
        - intents
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmIntentRequest'
      responses:
        '200':
          description: 确认结果（committed/rejected）与 resultRef。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ConfirmationIntent'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /projects/{projectId}/plans/{planId}/reopens:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: reopenPlan
      summary: 重开已验收计划（owner；保留历史）
      tags:
        - plans
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenRequest'
      responses:
        '200':
          description: 重开后的计划。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Plan'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /projects/{projectId}/tasks/{taskId}/reopens:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: reopenTask
      summary: 重开已验收任务（reviewer；引用原验收）
      tags:
        - tasks
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenRequest'
      responses:
        '200':
          description: 重开结果。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  schemas:
...
    ConfirmationIntentRequest:
      type: object
      required:
        - operation
        - objectId
        - reviewHash
      properties:
        operation:
          type: string
          enum:
            - proposal.decision
            - task.acceptance
            - task.reopen
            - plan.acceptance
            - plan.reopen
            - handoff.send
            - handoff.decision
            - project.create
            - workflow.publish
        objectId:
          type: string
          format: uuid
        reviewHash:
          type: string
        payload:
          type: object
          additionalProperties: true
          description: 该 operation 的命令载荷快照；服务端计算规范化哈希。
    ConfirmationIntent:
      type: object
      required:
        - id
        - state
        - operation
        - objectId
        - expiresAt
      properties:
        id:
          type: string
          format: uuid
        state:
          type: string
          enum:
            - pending
            - committed
            - rejected
            - expired
            - stale
        operation:
          type: string
        objectId:
          type: string
          format: uuid
        requestedBy:
          type: object
          required:
            - userId
            - deviceName
          properties:
            userId:
              type: string
              format: uuid
            deviceName:
              type:
                - string
                - 'null'
        confirmUrl:
          type:
            - string
            - 'null'
          format: uri
        resultRef:
          type:
            - string
            - 'null'
        expiresAt:
          type: string
          format: date-time
    ConfirmIntentRequest:
      type: object
      required:
        - intentHash
        - decision
      properties:
        intentHash:
          type: string
        decision:
          type: string
          enum:
            - approve
            - reject
        csrfToken:
          type: string
...
    RequestMeta:
      type: object
      required:
        - requestId
        - serverTime
      properties:
        requestId:
          type: string
        serverTime:
          type: string
          format: date-time
        eventCursor:
          type: integer
          format: int64
          description: 当前项目事件序列游标，仅项目资源响应返回。
    AffectedResource:
      type: object
      required:
        - type
        - id
        - version
      properties:
        type:
          type: string
        id:
          type: string
        version:
          type: integer
          minimum: 1
    CommandMeta:
      allOf:
        - $ref: '#/components/schemas/RequestMeta'
        - type: object
          properties:
            affected:
              type: array
              items:
                $ref: '#/components/schemas/AffectedResource'
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 兜底文案；客户端按 code+details 映射，不解析句子。
        details:
          description: 结构化错误详情，按 code 不同而不同。
        retryable:
          type: boolean
          default: false
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNSUPPORTED_SCHEMA
        - INVALID_REFERENCE
        - EMAIL_IN_USE
        - UNAUTHENTICATED
        - SESSION_EXPIRED
        - GRANT_REVOKED
        - FORBIDDEN
        - HUMAN_CONFIRMATION_REQUIRED
        - NOT_FOUND
        - VERSION_CONFLICT
        - REVIEW_STALE
        - DEPENDENCY_CHANGED
        - INVALID_TRANSITION
        - IDEMPOTENCY_CONFLICT
        - SOURCE_VERSION_CONFLICT
        - EVENT_CURSOR_EXPIRED
        - REQUIREMENT_UNMET
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA
        - RATE_LIMITED
        - BUDGET_EXHAUSTED
        - DEPENDENCY_UNAVAILABLE
        - MODEL_UNAVAILABLE
        - SANDBOX_UNAVAILABLE
        - INTERNAL_ERROR
    ErrorResponse:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        requestId:
          type: string
...
    Plan:
      type: object
      required:
        - id
        - title
        - status
        - ownerIdentityId
        - workflowId
        - version
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        goal:
          type: string
        acceptanceCriteria:
          type: string
        status:
          type: string
          enum:
            - draft
            - active
            - accepted
            - cancelled
        ownerIdentityId:
          type:
            - string
            - 'null'
          format: uuid
        workflowId:
          type:
            - string
            - 'null'
          format: uuid
        latestAcceptanceId:
          type:
            - string
            - 'null'
          format: uuid
        taskStats:
          type: object
          required:
            - total
            - accepted
            - active
            - cancelled
          properties:
            total:
              type: integer
            accepted:
              type: integer
            active:
              type: integer
            cancelled:
              type: integer
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
...
    Task:
      type: object
      required:
        - id
        - title
        - kind
        - status
        - reviewerIdentityId
        - version
      properties:
        id:
          type: string
          format: uuid
        planId:
          type:
            - string
            - 'null'
          format: uuid
        parentTaskId:
          type:
            - string
            - 'null'
          format: uuid
        title:
          type: string
          maxLength: 200
        expectedOutput:
          type: string
        acceptanceCriteria:
          type: string
        kind:
          type: string
          enum:
            - task
            - bug
        status:
          type: string
          enum:
            - draft
            - ready
            - working
            - delivered
            - accepted
            - rework
            - cancelled
        participants:
          type: array
          items:
            type: object
            required:
              - identityId
              - responsibility
            properties:
              identityId:
                type: string
                format: uuid
              displayName:
                type: string
              responsibility:
                type: string
        reviewerIdentityId:
          type:
            - string
            - 'null'
          format: uuid
        workflowId:
          type:
            - string
            - 'null'
          format: uuid
        nodeId:
          type:
            - string
            - 'null'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/TaskRequirement'
        latestReportId:
          type:
            - string
            - 'null'
          format: uuid
        latestAcceptanceId:
          type:
            - string
            - 'null'
          format: uuid
        blockers:
          type: array
          description: 当前阻塞当前置（由服务端计算）。
          items:
            type: object
            required:
              - phase
              - objectType
              - objectId
              - reason
            properties:
              phase:
                type: string
                enum:
                  - start
                  - accept
              objectType:
                type: string
              objectId:
                type: string
              reason:
                type: string
        version:
          type: integer
          minimum: 1
    TaskRequirement:
      type: object
      required:
        - id
        - phase
        - kind
        - targetId
        - hard
      properties:
        id:
          type: string
          format: uuid
        phase:
          type: string
          enum:
            - start
            - accept
            - both
        kind:
          type: string
          enum:
            - task_acceptance
            - handoff_receipt
            - material_ready
        targetId:
          type: string
          format: uuid
        materialVersionId:
          type:
            - string
            - 'null'
          format: uuid
        hard:
          type: boolean
        label:
          type: string
        satisfied:
          type: boolean
...
    ReopenRequest:
      type: object
      required:
        - expectedVersion
        - reason
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        acceptanceId:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 2000
...
  parameters:
...
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端生成的 UUID；同键同载荷重试返回原结果，同键异载荷返回 409。
      schema:
        type: string
        format: uuid
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: 未认证（401，UNAUTHENTICATED/SESSION_EXPIRED/GRANT_REVOKED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 已认证但无权限（403，FORBIDDEN/HUMAN_CONFIRMATION_REQUIRED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或对非成员隐藏存在性（404）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 状态/版本/幂等冲突（409，详见 error.code）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-judex_session
      description: Web 会话 Cookie（本地 HTTP 开发使用 judex_dev_session）。
    bearerGrant:
      type: http
      scheme: bearer
      description: CLI access token（受限 grant，无人工审批能力）。